+++ src/main/java/ru/practicum/shareit/user/User.java
@@ -2,12 +2,9 @@ package ru.practicum.shareit.user;
 
 import lombok.Data;
 
-/**
- * TODO Sprint add-controllers.
- */
 @Data
 public class User {
-    Integer id;
+    Long id;
     String name;
     String email;
 }
